use longs instead of ints for length

+++ trunk/src/audio/datasource_filereader.h
@@ -13,7 +13,7 @@ class FileReaderDataSource: public IDataSource
 		void reset();
 		bool exhausted();
 		long getpos();
-		int read(char* const buffer, unsigned long len);
+		unsigned long read(char* const buffer, unsigned long len);
 	private:
 		FILE* FileHandle;
 };
